+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_platform_interface/lib/kevin_flutter_accounts_platform_interface.dart
@@ -1,4 +1,8 @@
 library kevin_flutter_accounts_platform_interface;
 
+export 'package:kevin_flutter_core_platform_interface/src/error/kevin_error_helper.dart';
+export 'package:kevin_flutter_core_platform_interface/src/model/kevin_core_models.dart';
+
 export 'src/kevin_flutter_accounts_platform_interface.dart';
 export 'src/model/kevin_accounts_models.dart';
+export 'src/model/kevin_flutter_accounts_methods.dart';