+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_platform_interface/lib/src/model/account/kevin_account_linking_type.dart
@@ -7,9 +7,9 @@ enum KevinAccountLinkingType {
 }
 
 extension _FirstWhereOrNullExtension<E> on Iterable<E> {
-  E? firstWhereOrNull(bool Function(E) test) {
+  E? firstWhereOrNull(bool Function(E) predicate) {
     for (var element in this) {
-      if (test(element)) return element;
+      if (predicate(element)) return element;
     }
     return null;
   }