+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_platform_interface/lib/src/model/kevin_bank.dart
@@ -12,4 +12,24 @@ class KevinBank {
     required this.imageUri,
     required this.bic,
   });
+
+  Map<String, dynamic> toMap() {
+    return {
+      'id': id,
+      'name': name,
+      'officialName': officialName,
+      'imageUri': imageUri,
+      'bic': bic,
+    };
+  }
+
+  factory KevinBank.fromMap(Map<String, dynamic> map) {
+    return KevinBank(
+      id: map['id'] as String,
+      name: map['name'] as String,
+      officialName: map['officialName'] as String?,
+      imageUri: map['imageUri'] as String,
+      bic: map['bic'] as String?,
+    );
+  }
 }