+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_platform_interface/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.2
+
+* **BREAKING**: rename disabledCountrySelection to disableCountrySelection in
+  KevinAccountSessionConfiguration
+* **BREAKING**: use KevinCallbackUrl in KevinAccountsConfiguration instead of string
+* Upgrade kevin_flutter_core to 1.0.3
+* Add toMap/fromMap() methods to models where needed
+* Introduce KevinFlutterAccountsMethods for internal usage
+* Remove entity classes
+
 ## 1.0.1
 
 * Fix KevinAccountSessionConfigurationEntity.fromModel() method